requirements.txt
google-adk>=0.1.0
google-cloud-aiplatform>=1.60.0
google-generativeai>=0.7.0
braintrust>=0.0.100
datadog-api-client>=2.25.0
ddtrace>=2.8.0
elevenlabs>=1.1.2
python-dotenv>=1.0.0
requests>=2.31.0
python-nmap>=0.7.1
pyyaml>=6.0.1
rich>=13.7.1
asyncio
aiohttp>=3.9.0
